add fireball attack for boss, fix cutscene

+++ Assets/Scripts/Boss/BossHealthBar.cs
@@ -5,18 +5,18 @@ public class BossHealthBar : MonoBehaviour
 {
     public Image lightBar;
 
-    private int totalSegments;
+    private float totalSegments;
     private float targetFillAmount;
     private float lerpSpeed = 5f;
 
-    public void Setup(int total)
+    public void Setup(float total)
     {
         totalSegments = total;
         targetFillAmount = 1f;
         lightBar.fillAmount = 1f;
     }
 
-    public void SetHealth(int currentSegments)
+    public void SetHealth(float currentSegments)
     {
         if (totalSegments <= 0) return;
